There we go, a true alpha release. almost.
This update adds the main project file of the app. Newest features should allow creation of vertices with "newPoint (list 53 23 73)". this will in theory create a point at the position [53, 23, 73]. Will bugcheck tomorrow, four hours trying to get a straight answer on how to portray an isometric point on a 2 dimensional screen is way too draining. USE WITH CAUTION. Constructed with the community edition build, so there's bound to be inherent errors.