Derp. now it should be semi-functional.
the new format for the vertex feature is "newPoint (name) (x value) (y value) (z value)" (everything in parentheses is one word or number, you don't need to type x, y, or z. that's just for your benefit).
Bear in mind that the ViewPort is using an isometric viewing style by default. I will later update this to respect moving the camera around the origin, but it will STAY in Parallel Projection format because of the familiarity with diagrams for furniture and household materials. We don't need a perspective projection so there's no point in trying to break the bank coding one.
So the command "newPoint a 1 1 1" will not be separate from the result of the command "newPoint OOGABOOGAPUNK 42 42 42", they will be exactly the same place on the screen, despite one being about 71.01 arbitrary units away from the other.