Alpha 0.0.2
Lines are apparently more complicated than they look. But here's how to make them:
newLine a b ab true
"newLine" is the command. "a" is an existing vertex with the ID "a". Same with "b". "ab" is the name of the line. the "true" command is optional, if you put false it will in the
future allow you to curve the line, but for now, it will probably try to scream at you. Omitting it is perfectly legal though.

Not much else, however, the Line system is a temporary fix until I can get how I wanted it to work, to work. Hey, at least it's not Malbolge. That would suck.

I also added cartesian axes. You're welcome to adjust the Viewfinder class to make the axes thicker or thinner, just be warned that you should probably clone the file as a backup in case something gets irreparably screwed up.